refactor: move common types and schema
