+++ src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
 import { useAppSelector } from '../store/hooks';
-import { ModifiedFormData } from '../store/formSlice';
+import { ModifiedFormData } from '../types/common';
 
 const MainPage = () => {
   const { form } = useAppSelector((state) => state.form);
@@ -10,7 +10,13 @@ const MainPage = () => {
       <ul>
         {form.map((item: ModifiedFormData, idx) => (
           <li key={idx}>
-            <img src={item.userImage} style={{objectFit: 'cover'}} width='200px' height='200px' alt={item.name} />
+            <img
+              src={item.userImage}
+              style={{ objectFit: 'cover' }}
+              width="200px"
+              height="200px"
+              alt={item.name}
+            />
             {item.name}, {item.age}, {item.country}, {item.email}, {item.gender}
           </li>
         ))}
