+++ src/store/formSlice.tsx
@@ -1,20 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-export interface ModifiedFormData {
-  name: string;
-  age: number;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  gender: string;
-  acceptTerms?: boolean;
-  userImage?: string;
-  country: string;
-}
-
-interface FormsData {
-  form: ModifiedFormData[];
-}
+import { ModifiedFormData, FormsData } from '../types/common';
 
 const initialState: FormsData = {
   form: [],
